requirements.txt
# Core dependencies (Phase 1 & 2)
groq>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Phase 2 - Database
supabase>=2.0.0

# Phase 2 - Job scraping
# python-jobspy>=1.1.36  # Keep commented due to numpy conflicts - enable when needed

# Phase 2 - Google API (for Docs creation)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Testing
pytest>=7.4.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
